Refactor for KopfObject
