+++ operator/resourceclaim.py
@@ -9,6 +9,7 @@ from typing import List, Mapping, Optional, TypeVar, Union
 
 from deep_merge import deep_merge
 from jsonpatch_from_diff import jsonpatch_from_diff
+from kopfobject import KopfObject
 from poolboy import Poolboy
 from poolboy_templating import recursive_process_template_strings
 
@@ -20,9 +21,14 @@ ResourceClaimT = TypeVar('ResourceClaimT', bound='ResourceClaim')
 ResourceHandleT = TypeVar('ResourceHandleT', bound='ResourceHandle')
 ResourceProviderT = TypeVar('ResourceProviderT', bound='ResourceProvider')
 
-class ResourceClaim:
+class ResourceClaim(KopfObject):
+    api_group = Poolboy.operator_domain
+    api_version = Poolboy.operator_version
+    kind = "ResourceClaim"
+    plural = "resourceclaims"
+
     instances = {}
-    lock = asyncio.Lock()
+    class_lock = asyncio.Lock()
 
     @classmethod
     def __register_definition(cls, definition: Mapping) -> ResourceClaimT:
@@ -47,7 +53,7 @@ class ResourceClaim:
 
     @classmethod
     async def get(cls, name: str, namespace: str) -> ResourceClaimT:
-        async with cls.lock:
+        async with cls.class_lock:
             resource_claim = cls.instances.get((namespace, name))
             if resource_claim:
                 return resource_claim
@@ -68,7 +74,7 @@ class ResourceClaim:
         status: kopf.Status,
         uid: str,
     ) -> ResourceClaimT:
-        async with cls.lock:
+        async with cls.class_lock:
             resource_claim = cls.instances.get((namespace, name))
             if resource_claim:
                 resource_claim.refresh(
@@ -98,36 +104,13 @@ class ResourceClaim:
         cls,
         definition: Mapping,
     ) -> ResourceClaimT:
-        async with cls.lock:
+        async with cls.class_lock:
             return cls.__register_definition(definition=definition)
 
     @classmethod
     async def unregister(cls, name: str, namespace: str) -> Optional[ResourceClaimT]:
-        async with cls.lock:
+        async with cls.class_lock:
             return cls.instances.pop((namespace, name), None)
-
-    def __init__(self,
-        annotations: Union[kopf.Annotations, Mapping],
-        labels: Union[kopf.Labels, Mapping],
-        meta: Union[kopf.Meta, Mapping],
-        name: str,
-        namespace: str,
-        spec: Union[kopf.Spec, Mapping],
-        status: Union[kopf.Status, Mapping],
-        uid: str,
-    ):
-        self.annotations = annotations
-        self.labels = labels
-        self.lock = asyncio.Lock()
-        self.meta = meta
-        self.name = name
-        self.namespace = namespace
-        self.spec = spec
-        self.status = status
-        self.uid = uid
-
-    def __str__(self) -> str:
-        return f"ResourceClaim {self.name} in {self.namespace}"
 
     @property
     def approval_state(self) -> Optional[str]:
@@ -147,14 +130,6 @@ class ResourceClaim:
     @property
     def claim_is_initialized(self) -> bool:
         return f"{Poolboy.operator_domain}/resource-claim-init-timestamp" in self.annotations
-
-    @property
-    def creation_datetime(self):
-        return datetime.strptime(self.creation_timestamp, "%Y-%m-%dT%H:%H:%S%z")
-
-    @property
-    def creation_timestamp(self) -> str:
-        return self.meta['creationTimestamp']
 
     @property
     def has_resource_handle(self) -> bool:
@@ -242,10 +217,6 @@ class ResourceClaim:
         if timestamp:
             return timestamp
         return self.creation_timestamp
-
-    @property
-    def metadata(self) -> Mapping:
-        return self.meta
 
     @property
     def parameter_values(self) -> Mapping:
@@ -438,29 +409,6 @@ class ResourceClaim:
             return None
         return self.status['resources'][resource_number].get('state')
 
-    def refresh(self,
-        annotations: kopf.Annotations,
-        labels: kopf.Labels,
-        meta: kopf.Meta,
-        spec: kopf.Spec,
-        status: kopf.Status,
-        uid: str,
-    ) -> None:
-        self.annotations = annotations
-        self.labels = labels
-        self.meta = meta
-        self.spec = spec
-        self.status = status
-        self.uid = uid
-
-    def refresh_from_definition(self, definition: Mapping) -> None:
-        self.annotations = definition['metadata'].get('annotations', {})
-        self.labels = definition['metadata'].get('labels', {})
-        self.meta = definition['metadata']
-        self.spec = definition['spec']
-        self.status = definition.get('status', {})
-        self.uid = definition['metadata']['uid']
-
     async def update_status_from_handle(self,
         logger: kopf.ObjectLogger,
         resource_handle: ResourceHandleT
@@ -592,15 +540,6 @@ class ResourceClaim:
             f"to ResourceClaim {self.name} in {self.namespace}"
         )
 
-    async def delete(self):
-        await Poolboy.custom_objects_api.delete_namespaced_custom_object(
-            group = Poolboy.operator_domain,
-            name = self.name,
-            namespace = self.namespace,
-            plural = 'resourceclaims',
-            version = Poolboy.operator_version,
-        )
-
     async def detach(self, resource_handle):
         await self.merge_patch_status({
             "resourceHandle": {
@@ -665,19 +604,6 @@ class ResourceClaim:
 
         await self.merge_patch(patch)
         logger.info(f"ResourceClaim {self.name} in {self.namespace} initialized")
-
-    async def json_patch_status(self, patch: List[Mapping]) -> None:
-        """Apply json patch to object status and update definition."""
-        definition = await Poolboy.custom_objects_api.patch_namespaced_custom_object_status(
-            group = Poolboy.operator_domain,
-            name = self.name,
-            namespace = self.namespace,
-            plural = 'resourceclaims',
-            version = Poolboy.operator_version,
-            body = patch,
-            _content_type = 'application/json-patch+json',
-        )
-        self.refresh_from_definition(definition)
 
     async def manage(self, logger) -> None:
         async with self.lock:
@@ -881,44 +807,7 @@ class ResourceClaim:
                 })
 
         if patch:
-            definition = await Poolboy.custom_objects_api.patch_namespaced_custom_object(
-                body = patch,
-                group = Poolboy.operator_domain,
-                name = resource_handle.name,
-                namespace = resource_handle.namespace,
-                plural = 'resourcehandles',
-                version = Poolboy.operator_version,
-                _content_type = 'application/json-patch+json',
-            )
-            resource_handle.refresh_from_definition(definition)
-
-    async def merge_patch(self, patch: Mapping) -> None:
-        """Apply merge patch to object status and update definition."""
-        definition = await Poolboy.custom_objects_api.patch_namespaced_custom_object(
-            group = Poolboy.operator_domain,
-            name = self.name,
-            namespace = self.namespace,
-            plural = 'resourceclaims',
-            version = Poolboy.operator_version,
-            body = patch,
-            _content_type = 'application/merge-patch+json'
-        )
-        self.refresh_from_definition(definition)
-
-    async def merge_patch_status(self, patch: Mapping) -> None:
-        """Apply merge patch to object status and update definition."""
-        definition = await Poolboy.custom_objects_api.patch_namespaced_custom_object_status(
-            group = Poolboy.operator_domain,
-            name = self.name,
-            namespace = self.namespace,
-            plural = 'resourceclaims',
-            version = Poolboy.operator_version,
-            body = {
-                "status": patch
-            },
-            _content_type = 'application/merge-patch+json'
-        )
-        self.refresh_from_definition(definition)
+            await resource_handle.json_patch(patch)
 
     async def refetch(self) -> Optional[ResourceClaimT]:
         try:
